remove unnecessary dynamic import
